requirements.txt
colorama>=0.4.6
psutil>=5.9.5
PyYAML>=6.0
databases[sqlite]>=0.7.0
disnake>=2.9.0
asyncpraw>=7.7.0